further attempt to remove need for 'raw'
