+++ server/proxy-request.js
@@ -1,7 +1,6 @@
 /* jshint node:true */
 var requestp = require('./requestp');
 var u = require('underscore');
-var JSON3 = require('json3');
 var cloudflare = require('./cloudflare');
 
 
@@ -22,38 +21,14 @@ var passThroughHeaders = function(incomingHeaders) {
   return externalReqHeaders;
 };
 
-var isValidJson = function(str) {
-  try {
-    JSON3.parse(str);
-  } catch (e) {
-    return false;
-  }
-  return true;
-};
 
-
-module.exports = function(url, headers, raw) {
+module.exports = function(url, headers) {
   var externalReqHeaders = passThroughHeaders(headers);
 
-  var promise = requestp({
+  return requestp({
     uri: url,
     strictSSL: false, // node(jitsu?) has some SSL problems
     headers: externalReqHeaders,
     encoding: 'utf8'
-  }).then(
-    // process API response
-    function(response){
-      var body = response.body;
-
-      if (!raw && !isValidJson(body)){
-        // invalid JSON
-        throw new Error(body);
-      } else {
-        // proxy successful
-        return response;
-      }
-    }
-  );
-
-  return promise;
+  });
 };
